[components] extract bounding box animations
