+++ portfolio/components/ui/Figures/PhotoReceiptBoundingBox.tsx
@@ -4,8 +4,17 @@ import {
   type Line,
   type Point as ApiPoint,
 } from "../../../types/api";
-import { useSpring, useTransition, animated } from "@react-spring/web";
+import { useTransition, animated } from "@react-spring/web";
 import AnimatedLineBox from "../animations/AnimatedLineBox";
+import {
+  AnimatedConvexHull,
+  AnimatedHullCentroid,
+  AnimatedOrientedAxes,
+  AnimatedPrimaryEdges,
+  AnimatedSecondaryBoundaryLines,
+  AnimatedPrimaryBoundaryLines,
+  AnimatedReceiptFromHull
+} from "../animations";
 import { getBestImageUrl } from "../../../utils/imageFormat";
 import useImageDetails from "../../../hooks/useImageDetails";
 import {
@@ -235,843 +244,6 @@ const findLineEdgesAtPrimaryExtremes = (
  * Find the top and bottom edges of lines at the secondary axis extremes
  */
 
-// AnimatedConvexHull: component for animating convex hull calculation
-interface AnimatedConvexHullProps {
-  hullPoints: Point[];
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedConvexHull: React.FC<AnimatedConvexHullProps> = ({
-  hullPoints,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  const [visiblePoints, setVisiblePoints] = useState(0);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      const interval = setInterval(() => {
-        setVisiblePoints((prev) => {
-          if (prev >= hullPoints.length) {
-            clearInterval(interval);
-            return prev;
-          }
-          return prev + 1;
-        });
-      }, 200); // Add a new point every 200ms
-
-      return () => clearInterval(interval);
-    }, delay);
-
-    return () => clearTimeout(timer);
-  }, [delay, hullPoints.length]);
-
-  // Reset when hullPoints change (for animation reset)
-  useEffect(() => {
-    setVisiblePoints(0);
-  }, [hullPoints]);
-
-  if (hullPoints.length === 0) return null;
-
-  // Convert normalized coordinates to SVG coordinates
-  const svgPoints = hullPoints.map((point) => ({
-    x: point.x * svgWidth,
-    y: (1 - point.y) * svgHeight,
-  }));
-
-  // Create path for the visible portion of the hull
-  const visibleSvgPoints = svgPoints.slice(0, visiblePoints);
-
-  if (visibleSvgPoints.length < 2) {
-    return (
-      <>
-        {visibleSvgPoints.map((point, index) => (
-          <circle
-            key={index}
-            cx={point.x}
-            cy={point.y}
-            r={12}
-            fill="var(--color-red)"
-            opacity={1}
-            strokeWidth="2"
-          />
-        ))}
-      </>
-    );
-  }
-
-  const pathData = visibleSvgPoints.reduce((acc, point, index) => {
-    if (index === 0) return `M ${point.x} ${point.y}`;
-    return `${acc} L ${point.x} ${point.y}`;
-  }, "");
-
-  // Close the path if we've shown all points
-  const finalPath =
-    visiblePoints >= hullPoints.length ? `${pathData} Z` : pathData;
-
-  return (
-    <>
-      {/* Hull vertices */}
-      {visibleSvgPoints.map((point, index) => (
-        <circle
-          key={index}
-          cx={point.x}
-          cy={point.y}
-          r={12}
-          fill="var(--color-red)"
-          opacity={1}
-          strokeWidth="2"
-        />
-      ))}
-      {/* Hull edges */}
-      <path
-        d={finalPath}
-        fill="none"
-        stroke="var(--color-red)"
-        strokeWidth="4"
-        opacity={1}
-      />
-      {/* Fill when complete */}
-      {visiblePoints >= hullPoints.length && (
-        <path d={finalPath} fill="var(--color-red)" opacity={0.1} />
-      )}
-    </>
-  );
-};
-
-// AnimatedHullCentroid: component for visualizing hull centroid calculation
-interface AnimatedHullCentroidProps {
-  centroid: Point;
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedHullCentroid: React.FC<AnimatedHullCentroidProps> = ({
-  centroid,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  const centroidSpring = useSpring({
-    from: { opacity: 0, scale: 0 },
-    to: { opacity: 1, scale: 1 },
-    delay: delay,
-    config: { duration: 600 },
-  });
-
-  const centroidX = centroid.x * svgWidth;
-  const centroidY = (1 - centroid.y) * svgHeight;
-
-  return (
-    <animated.circle
-      cx={centroidX}
-      cy={centroidY}
-      r={15}
-      fill="var(--color-red)"
-      strokeWidth="3"
-      style={{
-        opacity: centroidSpring.opacity,
-        transform: centroidSpring.scale.to((s) => `scale(${s})`),
-        transformOrigin: `${centroidX}px ${centroidY}px`,
-      }}
-    />
-  );
-};
-
-// AnimatedOrientedAxes: component for visualizing oriented axes based on average line angle
-interface AnimatedOrientedAxesProps {
-  hull: Point[];
-  centroid: Point;
-  lines: Line[];
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedOrientedAxes: React.FC<AnimatedOrientedAxesProps> = ({
-  hull,
-  centroid,
-  lines,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  const [visibleElements, setVisibleElements] = useState(0);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      const interval = setInterval(() => {
-        setVisibleElements((prev) => {
-          if (prev >= 3) {
-            // 0: primary axis, 1: secondary axis, 2: extent points
-            clearInterval(interval);
-            return prev;
-          }
-          return prev + 1;
-        });
-      }, 400); // Show each element every 400ms
-
-      return () => clearInterval(interval);
-    }, delay);
-
-    return () => clearTimeout(timer);
-  }, [delay]);
-
-  // Reset when hull changes
-  useEffect(() => {
-    setVisibleElements(0);
-  }, [hull, lines]);
-
-  if (hull.length === 0 || lines.length === 0) return null;
-
-  // 1) Derive raw angles from each line’s bottom edge
-  const computedAngles = lines
-    .map((line) => {
-      const dx = line.bottom_right.x - line.bottom_left.x;
-      const dy = line.bottom_right.y - line.bottom_left.y;
-      // angle in degrees, 0 = perfectly horizontal
-      return (Math.atan2(dy, dx) * 180) / Math.PI;
-    })
-    // 2) Discard anything within ±0.001° of horizontal
-    .filter((angle) => Math.abs(angle) > 1e-3);
-
-  // 3) Average the rest, or fallback to 0 if none remain
-  const avgAngle =
-    computedAngles.length > 0
-      ? computedAngles.reduce((sum, a) => sum + a, 0) / computedAngles.length
-      : 0;
-
-  const centroidX = centroid.x * svgWidth;
-  const centroidY = (1 - centroid.y) * svgHeight;
-
-  // Convert angle to radians for calculations
-  const angleRad = (avgAngle * Math.PI) / 180;
-  const primaryAxisAngle = angleRad;
-  const secondaryAxisAngle = angleRad + Math.PI / 2; // Perpendicular axis (no longer used)
-
-  // Calculate axis endpoints (fixed 200px length for both axes)
-  const axisLength = 200;
-
-  // Primary axis (along average line direction)
-  const primaryAxis = {
-    x1: centroidX,
-    y1: centroidY,
-    x2: centroidX + axisLength * Math.cos(primaryAxisAngle),
-    y2: centroidY + axisLength * Math.sin(primaryAxisAngle),
-  };
-
-  // Secondary axis always pointing straight up from the centroid
-  const secondaryAxis = {
-    x1: centroidX,
-    y1: centroidY,
-    x2: centroidX,
-    y2: centroidY - axisLength,
-  };
-
-  // Find extent points along each axis
-  let minPrimary = Infinity,
-    maxPrimary = -Infinity;
-  let minSecondary = Infinity,
-    maxSecondary = -Infinity;
-  let primaryMinPoint = hull[0],
-    primaryMaxPoint = hull[0];
-  let secondaryMinPoint = hull[0],
-    secondaryMaxPoint = hull[0];
-
-  // Calculate projections for all hull points
-  const hullProjections = hull.map((point) => {
-    const px = point.x * svgWidth;
-    const py = (1 - point.y) * svgHeight;
-    const relX = px - centroidX;
-    const relY = py - centroidY;
-    const primaryProjection =
-      relX * Math.cos(primaryAxisAngle) + relY * Math.sin(primaryAxisAngle);
-    const secondaryProjection =
-      relX * Math.cos(secondaryAxisAngle) + relY * Math.sin(secondaryAxisAngle);
-
-    return {
-      point: { x: px, y: py },
-      primaryProjection,
-      secondaryProjection,
-      originalPoint: point,
-    };
-  });
-
-  // Find the extreme projections
-  hullProjections.forEach(
-    ({ point, primaryProjection, secondaryProjection }) => {
-      if (primaryProjection < minPrimary) {
-        minPrimary = primaryProjection;
-        primaryMinPoint = point;
-      }
-      if (primaryProjection > maxPrimary) {
-        maxPrimary = primaryProjection;
-        primaryMaxPoint = point;
-      }
-      if (secondaryProjection < minSecondary) {
-        minSecondary = secondaryProjection;
-        secondaryMinPoint = point;
-      }
-      if (secondaryProjection > maxSecondary) {
-        maxSecondary = secondaryProjection;
-        secondaryMaxPoint = point;
-      }
-    }
-  );
-
-  // Find the single closest hull point to each secondary extreme
-  let topExtremeClosest = secondaryMaxPoint;
-  let bottomExtremeClosest = secondaryMinPoint;
-  let minTopDistance = Infinity;
-  let minBottomDistance = Infinity;
-
-  hullProjections.forEach(({ point, secondaryProjection }) => {
-    // Skip if this is already the extreme point itself
-    if (
-      Math.abs(point.x - secondaryMaxPoint.x) < 1 &&
-      Math.abs(point.y - secondaryMaxPoint.y) < 1
-    )
-      return;
-    if (
-      Math.abs(point.x - secondaryMinPoint.x) < 1 &&
-      Math.abs(point.y - secondaryMinPoint.y) < 1
-    )
-      return;
-
-    // Find closest to top extreme
-    const topDistance = Math.abs(secondaryProjection - maxSecondary);
-    if (topDistance < minTopDistance) {
-      minTopDistance = topDistance;
-      topExtremeClosest = point;
-    }
-
-    // Find closest to bottom extreme
-    const bottomDistance = Math.abs(secondaryProjection - minSecondary);
-    if (bottomDistance < minBottomDistance) {
-      minBottomDistance = bottomDistance;
-      bottomExtremeClosest = point;
-    }
-  });
-
-  // Combine primary extremes (green)
-  const primaryPoints = [primaryMinPoint, primaryMaxPoint];
-  // secondaryMinPoint and secondaryMaxPoint are already in scope
-
-  return (
-    <>
-      <defs>
-        <marker
-          id="axis-arrow-primary"
-          markerWidth="8"
-          markerHeight="8"
-          refX="0"
-          refY="3"
-          orient="auto"
-          markerUnits="strokeWidth"
-        >
-          <path d="M0,0 L0,6 L6,3 Z" fill="var(--color-green)" />
-        </marker>
-        <marker
-          id="axis-arrow-secondary"
-          markerWidth="8"
-          markerHeight="8"
-          refX="0"
-          refY="3"
-          orient="auto"
-          markerUnits="strokeWidth"
-        >
-          <path d="M0,0 L0,6 L6,3 Z" fill="var(--color-yellow)" />
-        </marker>
-      </defs>
-      {/* Primary axis (along average line direction) */}
-      {visibleElements >= 1 && (
-        <line
-          x1={primaryAxis.x1}
-          y1={primaryAxis.y1}
-          x2={primaryAxis.x2}
-          y2={primaryAxis.y2}
-          stroke="var(--color-green)"
-          strokeWidth="10"
-          opacity={0.8}
-          markerEnd="url(#axis-arrow-primary)"
-        />
-      )}
-
-      {/* Secondary axis (perpendicular to average line direction) */}
-      {visibleElements >= 2 && (
-        <line
-          x1={secondaryAxis.x1}
-          y1={secondaryAxis.y1}
-          x2={secondaryAxis.x2}
-          y2={secondaryAxis.y2}
-          stroke="var(--color-yellow)"
-          strokeWidth="10"
-          opacity={0.8}
-          markerEnd="url(#axis-arrow-secondary)"
-        />
-      )}
-
-      {/* Primary extent points (green) */}
-      {visibleElements >= 3 &&
-        primaryPoints.map((point, index) => (
-          <circle
-            key={`primary-${index}`}
-            cx={point.x}
-            cy={point.y}
-            r={10}
-            fill="var(--color-green)"
-            stroke="white"
-            strokeWidth="2"
-            opacity={0.9}
-          />
-        ))}
-
-      {/* Hull left/right extremes (green) */}
-      {visibleElements >= 3 && (
-        <>
-          {[secondaryMinPoint, secondaryMaxPoint].map((point, idx) => (
-            <circle
-              key={`hull-secondary-extreme-${idx}`}
-              cx={point.x}
-              cy={point.y}
-              r={10}
-              fill="var(--color-green)"
-              stroke="white"
-              strokeWidth="2"
-              opacity={0.9}
-            />
-          ))}
-        </>
-      )}
-    </>
-  );
-};
-
-// AnimatedPrimaryEdges: show line edges at primary extremes
-interface AnimatedPrimaryEdgesProps {
-  lines: Line[];
-  hull: Point[];
-  centroid: Point;
-  avgAngle: number;
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedPrimaryEdges: React.FC<AnimatedPrimaryEdgesProps> = ({
-  lines,
-  hull,
-  centroid,
-  avgAngle,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  const { leftEdgePoints, rightEdgePoints } = findBoundaryLinesWithSkew(
-    lines,
-    hull,
-    centroid,
-    avgAngle
-  );
-
-  const allEdgePoints = [...leftEdgePoints, ...rightEdgePoints];
-
-  const edgeTransitions = useTransition(allEdgePoints, {
-    keys: (point) => `edge-${point.x}-${point.y}`,
-    from: { opacity: 0, scale: 0 },
-    enter: (item, index) => ({
-      opacity: 1,
-      scale: 1,
-      delay: delay + index * 100,
-    }),
-    config: { duration: 400 },
-  });
-
-  return (
-    <g>
-      {edgeTransitions((style, edgePoint) => {
-        const isLeftEdge = leftEdgePoints.some(
-          (p) => p.x === edgePoint.x && p.y === edgePoint.y
-        );
-        return (
-          <animated.circle
-            style={style}
-            cx={edgePoint.x * svgWidth}
-            cy={(1 - edgePoint.y) * svgHeight}
-            r="6"
-            fill={"var(--color-blue)"}
-            stroke="white"
-            strokeWidth="1"
-          />
-        );
-      })}
-    </g>
-  );
-};
-
-// AnimatedSecondaryBoundaryLines: draw lines through yellow circles (secondary extremes)
-interface AnimatedSecondaryBoundaryLinesProps {
-  lines: Line[];
-  hull: Point[];
-  centroid: Point;
-  avgAngle: number;
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-const AnimatedSecondaryBoundaryLines: React.FC<
-  AnimatedSecondaryBoundaryLinesProps
-> = ({ hull, centroid, avgAngle, svgWidth, svgHeight, delay }) => {
-  // Calculate data first to ensure consistent hook calls
-  let bottomPts: Point[] = [];
-  let topPts: Point[] = [];
-  let lineSegments: LineSegment[] = [];
-
-  if (hull.length >= 3) {
-    // 1) Compute perpendicular axis
-    const angleRad = (avgAngle * Math.PI) / 180;
-    const secondaryAxisAngle = angleRad + Math.PI / 2;
-    const cosS = Math.cos(secondaryAxisAngle);
-    const sinS = Math.sin(secondaryAxisAngle);
-
-    // 2) Project hull points, sort by projection
-    const projHull = hull
-      .map((p) => {
-        const rx = p.x - centroid.x;
-        const ry = p.y - centroid.y;
-        return { point: p, proj: rx * cosS + ry * sinS };
-      })
-      .sort((a, b) => a.proj - b.proj);
-
-    // 3) Grab the bottom‐two and top‐two extreme points
-    bottomPts = [projHull[0].point, projHull[1].point];
-    topPts = [
-      projHull[projHull.length - 2].point,
-      projHull[projHull.length - 1].point,
-    ];
-
-    // Helper to extend a segment between two hull points out to full SVG width
-    const extendFullWidth = (
-      pA: Point,
-      pB: Point,
-      key: string
-    ): LineSegment => {
-      const xA = pA.x * svgWidth,
-        yA = (1 - pA.y) * svgHeight;
-      const xB = pB.x * svgWidth,
-        yB = (1 - pB.y) * svgHeight;
-      const m = (yB - yA) / (xB - xA);
-      const c = yA - m * xA;
-      return {
-        key,
-        x1: 0,
-        y1: c,
-        x2: svgWidth,
-        y2: m * svgWidth + c,
-      };
-    };
-
-    // 4) Build two yellow line segments between each pair
-    lineSegments = [
-      extendFullWidth(bottomPts[0], bottomPts[1], "bottom-hull-boundary"),
-      extendFullWidth(topPts[0], topPts[1], "top-hull-boundary"),
-    ];
-  }
-
-  // Animate yellow dots at the four extreme hull points
-  const dotPoints = [...bottomPts, ...topPts];
-  const dotTransitions = useTransition(dotPoints, {
-    from: { opacity: 0 },
-    enter: (_pt, idx) => ({
-      opacity: 1,
-      delay: delay + idx * 200,
-    }),
-    config: { duration: 400 },
-  });
-
-  // 5) Animate exactly as before
-  const lineTransitions = useTransition(lineSegments, {
-    keys: (line) => line.key,
-    from: { opacity: 0, strokeDasharray: "10,10", strokeDashoffset: 20 },
-    enter: (_item, index) => ({
-      opacity: 1,
-      strokeDashoffset: 0,
-      delay: delay + index * 200,
-    }),
-    config: { duration: 800 },
-  });
-
-  if (hull.length < 3) return null;
-
-  return (
-    <g>
-      {/* Animated yellow dots */}
-      {dotTransitions((style, pt, _item, idx) => (
-        <animated.circle
-          key={`secondary-extreme-${idx}`}
-          style={style}
-          cx={pt.x * svgWidth}
-          cy={(1 - pt.y) * svgHeight}
-          r={8}
-          fill="var(--color-yellow)"
-          stroke="white"
-          strokeWidth="2"
-        />
-      ))}
-      {lineTransitions((style, seg) => (
-        <animated.line
-          key={seg.key}
-          style={style}
-          x1={seg.x1}
-          y1={seg.y1}
-          x2={seg.x2}
-          y2={seg.y2}
-          stroke="var(--color-yellow)"
-          strokeWidth="5"
-          strokeDasharray="10,10"
-        />
-      ))}
-    </g>
-  );
-};
-// AnimatedPrimaryBoundaryLines: draw left/right boundaries using hull edge binning and Theil–Sen fit
-interface AnimatedPrimaryBoundaryLinesProps {
-  hull: Point[];
-  centroid: Point;
-  avgAngle: number;
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedPrimaryBoundaryLines: React.FC<
-  AnimatedPrimaryBoundaryLinesProps
-> = ({ hull, centroid, avgAngle, svgWidth, svgHeight, delay }) => {
-  // Calculate segments first for proper hook usage
-  let segments: LineSegment[] = [];
-
-  if (hull.length >= 3) {
-    // 1) compute receipt tilt axis unit vector
-    const angleRad = (avgAngle * Math.PI) / 180;
-    const ux = Math.cos(angleRad),
-      uy = Math.sin(angleRad);
-
-    // 2) project hull points onto receipt axis
-    const projections = hull.map((p, i) => ({
-      idx: i,
-      proj: p.x * ux + p.y * uy,
-    }));
-    projections.sort((a, b) => a.proj - b.proj);
-
-    // 3) identify extremes along that axis
-    const extremes = [
-      { key: "left-boundary", idx: projections[0].idx },
-      { key: "right-boundary", idx: projections[projections.length - 1].idx },
-    ];
-
-    // 4) build segments using weighted cost comparison
-    const DIST_WEIGHT = 1;
-    const ANGLE_WEIGHT = 1;
-    const span = Math.hypot(svgWidth, svgHeight);
-
-    segments = extremes.map(({ key, idx }) => {
-      const p0 = hull[idx];
-      const pCW = hull[(idx + 1) % hull.length];
-      const pCCW = hull[(idx - 1 + hull.length) % hull.length];
-
-      // helper to fit line in ax+by+c=0
-      const makeLine = (p1: Point, p2: Point) => {
-        const a = p2.y - p1.y;
-        const b = p1.x - p2.x;
-        const c = p2.x * p1.y - p1.x * p2.y;
-        return { a, b, c };
-      };
-      const lineCW = makeLine(p0, pCW);
-      const lineCCW = makeLine(p0, pCCW);
-
-      // mean perpendicular-distance error
-      const meanDist = ({ a, b, c }: { a: number; b: number; c: number }) => {
-        const norm = Math.hypot(a, b);
-        return (
-          hull.reduce(
-            (sum, p) => sum + Math.abs(a * p.x + b * p.y + c) / norm,
-            0
-          ) / hull.length
-        );
-      };
-      const distCW = meanDist(lineCW);
-      const distCCW = meanDist(lineCCW);
-
-      // compute neighbor angles
-      const diffAngle = (p: Point) => {
-        const theta = Math.atan2(p.y - p0.y, p.x - p0.x);
-        const d = Math.abs(theta - angleRad);
-        return Math.min(d, 2 * Math.PI - d);
-      };
-      const diffCW = diffAngle(pCW);
-      const diffCCW = diffAngle(pCCW);
-
-      // weighted cost and pick
-      const costCW = DIST_WEIGHT * distCW + ANGLE_WEIGHT * diffCW;
-      const costCCW = DIST_WEIGHT * distCCW + ANGLE_WEIGHT * diffCCW;
-      const pB = costCW < costCCW ? pCW : pCCW;
-
-      // extend line full-height through p0->pB
-      const dx = (pB.x - p0.x) * svgWidth;
-      const dy = (pB.y - p0.y) * svgHeight;
-      const m = dx / dy;
-      const c = p0.x * svgWidth - m * (p0.y * svgHeight);
-
-      return {
-        key,
-        x1: c,
-        y1: 0,
-        x2: m * span + c,
-        y2: span,
-      };
-    });
-  }
-
-  // Hook must be called at the top level
-  const transitions = useTransition(segments, {
-    keys: (seg) => seg.key,
-    from: { opacity: 0, strokeDasharray: "12,8", strokeDashoffset: 20 },
-    enter: (_item, idx) => ({
-      opacity: 1,
-      strokeDashoffset: 0,
-      delay: delay + idx * 300,
-    }),
-    config: { duration: 800 },
-  });
-
-  if (hull.length < 3) return null;
-
-  return (
-    <g>
-      {segments.map((seg) => (
-        <React.Fragment key={`pts-${seg.key}`}>
-          <circle
-            cx={seg.x1}
-            cy={seg.y1}
-            r={8}
-            fill="var(--color-green)"
-            stroke="white"
-            strokeWidth="2"
-          />
-          <circle
-            cx={seg.x2}
-            cy={seg.y2}
-            r={8}
-            fill="var(--color-green)"
-            stroke="white"
-            strokeWidth="2"
-          />
-        </React.Fragment>
-      ))}
-      {transitions((style, seg) => (
-        <animated.line
-          key={seg.key}
-          style={style}
-          x1={seg.x1}
-          y1={seg.y1}
-          x2={seg.x2}
-          y2={seg.y2}
-          stroke="var(--color-green)"
-          strokeWidth="5"
-          strokeDasharray="12,8"
-        />
-      ))}
-    </g>
-  );
-};
-
-// AnimatedReceiptFromHull: component using the proper Python algorithm
-interface AnimatedReceiptFromHullProps {
-  hull: Point[];
-  lines: Line[];
-  svgWidth: number;
-  svgHeight: number;
-  delay: number;
-}
-
-const AnimatedReceiptFromHull: React.FC<AnimatedReceiptFromHullProps> = ({
-  hull,
-  lines,
-  svgWidth,
-  svgHeight,
-  delay,
-}) => {
-  // Initialize hooks at the top level
-  const boxSpring = useSpring({
-    from: { opacity: 0 },
-    to: { opacity: 1 },
-    delay: delay,
-    config: { duration: 800 },
-  });
-
-  const centroidSpring = useSpring({
-    from: { opacity: 0 },
-    to: { opacity: 1 },
-    delay: delay + 400,
-    config: { duration: 600 },
-  });
-
-  if (hull.length === 0 || lines.length === 0) return null;
-
-  // Compute hull centroid
-  const hullCentroid = computeHullCentroid(hull);
-
-  console.log(`lines[0]`);
-  // Use the actual angle data from OCR API (more accurate than manual calculation)
-  const avgAngle =
-    lines.reduce((sum, line) => sum + line.angle_degrees, 0) / lines.length;
-
-  // Compute receipt box using line edges at secondary extremes
-  const receiptCorners = computeReceiptBoxFromLineEdges(
-    lines,
-    hull,
-    hullCentroid,
-    avgAngle
-  );
-
-  if (receiptCorners.length !== 4) return null;
-
-  // Convert to SVG coordinates
-  const svgCorners = receiptCorners.map((corner) => ({
-    x: corner.x * svgWidth,
-    y: (1 - corner.y) * svgHeight,
-  }));
-
-  const points = svgCorners.map((c) => `${c.x},${c.y}`).join(" ");
-
-  // Compute receipt centroid for visualization
-  const receiptCentroidX = svgCorners.reduce((sum, c) => sum + c.x, 0) / 4;
-  const receiptCentroidY = svgCorners.reduce((sum, c) => sum + c.y, 0) / 4;
-
-  return (
-    <>
-      <animated.polygon
-        style={boxSpring}
-        points={points}
-        fill="none"
-        stroke="var(--color-blue)"
-        strokeWidth="4"
-      />
-      <animated.circle
-        cx={receiptCentroidX}
-        cy={receiptCentroidY}
-        r={12}
-        fill="var(--color-blue)"
-        strokeWidth="2"
-        style={{ opacity: centroidSpring.opacity }}
-      />
-    </>
-  );
-};
 
 /**
  * Display a random photo image with animated overlays that illustrate
